stupid change to avoide stupid eclipse indexing warnings
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5020 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/layers/Patterns.cpp
@@ -333,9 +333,9 @@ int Patterns::generatePattern(float val)
 
             //sin of 2*pi*m/wavelength, where wavelength=2*width:
             if((type == SINEWAVE)||(type == SINEV))
-               data[ix * sx + iy * sy] = sinf(PI*m/float(wavelength));
+               data[ix * sx + iy * sy] = sin(PI*m/float(wavelength));
             else if((type == COSWAVE)||(type == COSV))
-               data[ix * sx + iy * sy] = cosf(PI*m/float(wavelength));
+               data[ix * sx + iy * sy] = cos(PI*m/float(wavelength));
 
          }
       }
